refactor(gateway): name connector predicate module explicitly
