+++ src/handlers/gateway/connector/get/index.tsx
@@ -4,7 +4,7 @@ import { createHandler, flag } from "../../../../router";
 import { JsonRendererKey } from "../../../../tui";
 import type { Core } from "../../../types";
 import { coreOptsFromCtx } from "../../../utils";
-import { isConnectorTarget } from "../target";
+import { isConnectorTarget } from "../isConnectorTarget";
 
 export const createGetGatewayConnectorHandler = (core: Core) =>
   createHandler({
