+++ src/handlers/gateway/connector/isConnectorTarget.ts
